requirements.txt
fastapi
uvicorn
pydantic
pytest
pyjwt
authlib
